History: add decorator to record history events
The catmaid.history module now provides decorators and regular functions
to record history events for the current transaction. These provide a
convenient way to provide semantic information to the current
transaction. The transaction ID and time is stored in CATMAID's history
tables already.

Treenode creation is used as an example and will now record history
events.