Node: spatially query connectors and links
Previously, connectors were queried based on whether they were related
to a treenode in the bounding box or were in a z-section of the bounding
box. This had several downsides:

1. Joining a large set of treenodes into links when few would actually
   exist
2. Links where the connector and treenode were both outside the bounding
   box were omitted even if the link intersected the bounding box
3. Querying connectors by section axis does not generalize to arbitrary
   planes

Instead, query links and connectors spatially via a PostGIS edge table
as is done for treenodes.

There are several possible implementations. The one in this commit
creates separate geometry tables for treenode_connectors and connectors.

See catmaid/CATMAID#1085.