Node: spatially query connectors and links
Previously, connectors were queried based on whether they were related
to a treenode in the bounding box or were in a z-section of the bounding
box. This had several downsides:

1. Joining a large set of treenodes into links when few would actually
   exist
2. Links where the connector and treenode were both outside the bounding
   box were omitted even if the link intersected the bounding box
3. Querying connectors by section axis does not generalize to arbitrary
   planes

Instead, query links and connectors spatially via a PostGIS edge table
as is done for treenodes.

There are several possible implementations. The one in this commit
creates an edge table for treenode_connectors that:

1. Indexes based on project, treenode and connector, not
   treenode_connector's ID (which is not used elsewhere and is
   unindexed)
2. Deduplicates edges for multiple treenode_connectors for
   the same treenode and connector but different relations
3. Allows treenode to be NULL, so that connectors can also be added to
   this edge table like root nodes in the treenode_edge table

This increases complexity of the edge table maintenance triggers
substantially, but prevents the need to separately query for connectors
without links.