Check whether nodes are editable on client
Checking whether nodes are editable -- that is, whether the node creator
is in the request user's edit domain -- during node querying is wasteful
because:

- It centralizes work that could easily be done client side
- It does work in Python that could easily be done in faster JS
- It checks the editability of every queried node, even though this is
  orders of magnitude larger than the set of nodes on which edit
  attempts will be made
- This check is more meaningful as a client action, since it is only a
  UI cue (editability is *enforced* by the server separately)

Perform this check on the client by having the node query return nodes'
creating user ID instead of the 'can_edit' boolean. The only cost is the
potential for minutely larger node query responses (because IDs may be
larger than true/false and are less compressible), response parsing, and
node storage.