Set CATMAID session for anonymous users
Many tools and widgets depend on context available in CATMAID.session,
which was only set for logged in users. While in most setups anonymous
users would not be able to access these tools, that is not always the
case. For consistency, always set CATMAID.session, and rely on
CATMAID.session.id to distinguish non-anonymous sessions.

On the backend, factor out this context response for consistency, and
also return a full anonymous user context on logout. The client logout
now reuses the login code.