Neuron API: add endpoint to get neuron IDs based on model objects
There was no simple API to get all neuron IDs from a list of skeletons.
The /neuronname and /neuronnames endpoint return not only the ID, but
also neuron names and since the concept of neuron names is likely to
change soon, a new endpoint has been added to only retrieve the modeling
information. At the moment, only skeletons model neurons. This might
change in the future and therefore this endpoint is named more generic.

A unit test for this endpoint is included.