Tests: don't override anonymous user in test fixture
In a new test database the anonymous user is created through a
post_migrate signal handler and gets ID 1 assigned. This ID is
explicitly used for a user in our test fixture, which causes the
anonymous user to be overridden. To fix this, the test user with ID 1
now got the next free ID (5) assigned. Some tests had to be adapted for
this to work.

Fixes catmaid/CATMAID#1352