SVG overlay: only request state for node removal if active requests are done
Waiting for active requests to finish is important to be able to queue a delete
actions quickly after another state-changing action (concerning the node to
delete). For instance, quickly queuing multiple node removal actions wouldn't be
possible without this (because of state mismatch errors). Hitting delete very
quickly for a chain of nodes will still fail, because the active node is only
set to the parent after the child removal succeeds, but requested when delete is
hit.

See catmaid/CATMAID#1354