History tables: use original parameters for parent table tracking
In case history tracking is enabled for a parent table, the same
parameters should be used that were provided for the original target
table. This means triggers are only installed if they should be
installed on the original live table and the same is true for
inheritance checks and initial syncing. So far these parameters have
been forced to be true.