User stats: fix ORM problem with review stats query
So far the Django ORM was used to retrieve review statistics for a given
date range. For some reason, however, it "rounded" the end data in the
query to the next day. If this day contained reviews, rows for more
dates than expected were returned and the back-end code would break.
Cable length and connectors were retrieved with plain SQL already, so
they were not affected by this.

The review stats query is now replaced with a plain SQL version, which
fixes this problem and is slightly faster.

Also, the end date tests are now more precise and include all dates from
the passed in range end.