Tracing overlay: add visibility groups
Filter visibility of skeletons based on group membership. Group
membership is determined by meta-annotation or creator. Unlike skeleton
source subscription based visibility, these visibility groups can be
toggled quickly with HOME and SHIFT + HOME because the nodes are already
rendered in the tracing overlay.

This is an adaption of work done upstream for CATSOP.