History tables: add test to find new tables
If a new table is created in the database without marking it explicitly
as versioned or unversioned in this test, this test will raise an error.
With the help of this we should become aware of missing history tables
for newly created live tables (regardless of by us or one of our
dependencies). It also serves as a general test which tests our
expectations of tables in the database. For now, it only looks at tables
in the 'public' schema.