Tracing overlay: allow "G" on skeleton projections
Change the behavior of near-cursor selection so that if the node is in
the current section, the stack viewer is not moved (since this is
jarring), but if the nearest node is in another section, it is moved.

The restores the ability to select nodes from the skeleton projection
layer.