Transaction log: add project column and indices
The transaction log is accessible from the front-end, from where it
typically is only interested in transactions modifying the currently
open project. To make this sort of filtering quick, the project ID has
to be stored for each transaction entry. This column is of course
nullable. An index is added for this column.

Also an index is added for the execution time column in descending
order. This is useful for sorting transactions to e.g. show only recent
changes.