Ensure all projects have required classes and relations
With the help of a data migration all projects are checked if they have
all needed classes and relations available. If not, they are created. To
reuse code, the lists of required classes/relations have been moved out
of the model. Model fields can't accessed easily from withing
migrations. Therefore, these requirements are moved to control.project,
similar to tracing and classification dependencies.