History tracking: use tracking tables for all tables, also log txid
The original transaction ID (txid) of a row change is now stored as part
of a tracking table (previously time table) along with the creation time
stamp. Using a separate tracking table for all versioned tables is done
only for benchmarking. The next commit will make CATMAID tables keep
track of this themselves.

Up to now, the transaction ID of the live data modifying transaction was
stored in the history entry (i.e. a re-rooting operation is referenced
from history entries of the updated rows). This is now changed so that
the *original* transaction ID is stored that created/updated the old
live column. This ID is needed to properly relate history entries to old
transaction entries.

Even though the modifying transaction ID that caused the history change
is stored implicitly by Postgres as system column xmin, it is not safe
to use and rely on it (say to quickly find also all history rows created
by a particular transaction), because it gets overridden by transaction
ID wraparound with a frozen transaction ID.

Additionally, two more helper functions to reload history table triggers
are added: reload_history_tracking_for_table(regclass) and
reload_history_tracking(). These will re-create trigger functions for
the given table or all tables, respectively.