Improve robustness of RequestQueue's callback execution
In case of an exception, the queue is now always reset. This seems like
a safe default, because sending new requests in an erroneous state might
lead to unwanted result.

Not catching such exceptions would in the past lead to freezing of
CATMAID after an error dialog was closed. The problem being the
RequestQueue didn't shift after an error not was it reset. New requests
were waiting to be sent.

This might also help with catmaid/CATMAID#1287