Tracing overlay: efficiently create virtual nodes
Virtual nodes were made by checking, for each node not in the current
section, whether the parent or any children cross the section plane.
Because for each edge crossing the section plane we are guaranteed to
receive both the parent and the child, it is more efficient to only
check whether the edge to the parent crosses the section plane. This
avoids duplicate checks and duplicate creation of virtual nodes.