Request parent/child of virtual node for node query
When the active node is virtual, explicitly request that both the parent
and child be included in the backend node list query. To accomodate
this, change the node list query to take a list of treenode and
connector IDs to include in the result, rather than a single active node
ID and type.

Closes catmaid/CATMAID#1374.