Stack viewer: allow redraw calls to be blocking
Add a flag to the stack viewer which tells layers to make their redraws
blocking. That is, if the layer has an asynchronous component to
redrawing, do not invoke completion callbacks until the asynchronous
tasks are complete.

This is necessary for anything that depends on full completion of moving
and redraws before invoking its callback, such as animated scrolling.

Note that some layers may choose to be blocking even if this flag is not
set, like the tracing overlay.