Buffer geometries: add method to create Lambert material
Since buffer geometries themselves have a big influence on how the
material shader has to work, it makes sense to put the material
generation in the same class. This allows also for custom material
adaptions like providing alpha a uniform attribute to the shader if it
is the same for all.