Buffer geometries: add indexed multi object geometry
This new custom THREE geometry needs much less memory than the non-index
version. A template mesh is generated and instantiated on all the different
locations on the fly. This way only a single offset point is needed for
each instantiated mesh as spatial information. The greatly improves the
memory foot-print when many objects are in use (which is typically the
case with labels and connector nodes).

Note that bounding boxes and bounding spheres can not be computed using
THREE's computeBoundingBox() and computeBoundingSphere() method. The
geometry considers only the template vertices.