Treenode: improve node creation performance
Improve performance of node creation by using a faster query to get the
neuron ID modeled by the skeleton to which the treenode will belong.
Suprisingly this was usually the slowest part of node creation by an
order of magnitude. These changes reduce the time of this check by up to
10x at the cost of a new, more specialized permissions check.

This also makes _create_treenode more correct by locking the parent
before checking edit permissions, which avoids a race condition in the
permissions check. However, this condition could not occur in practice
because the parent is already locked by the state checking in the API
endpoint.