Buffer geometries: fix index mismatch after objects were skipped
During initialization of an instanced buffer geometry based on a list of
objects (e.g. connectors or labels) individual objects can be skipped
based on filter function. This has to be respected when finding the next
free index for new buffer object. Not doing so would lead to assigning
properties to the wrong object, visible for instance when changing the
shading mode or using picking.