History tracking: add combined live/history view for all versioned tables
Each versioned table now has a view associated that combines the content
of itself with the content of its history table. This is done in a view,
to not have the client take care of the columns and their ordering. They
are important to unify live and history data. This also means, history
views have to be regenerated if tables change.

History views have a __with_history suffix.