Fix performances test history tables before creating history views
Depending on the order in which the existing migrations have been
applied, an error was raised when applying the history view migration
of CATMAID. The problem was a performance test table (testview) where a
column data type was updated without also updating the table's history
table. The root of this problem is a missing dependency of the history
table migration: the JSONB data type upgrade should happen before the
history table migration. This way the history view migration would have
seen a consistent live/history table pair. This is now fixed by making
sure the JSONB data type is used in the performance test testview
history table and making this change a dependency of the history view
migration. This migration is aware of a potential manual fix of this
situation and only tries to change the data type of this history table
column if it hasn't the correct type already.