3D viewer: don't depend on tracing overlay for active node update
Users reported that the changed radius lookup would occasionaly fail, because
the active node wouldn't be available from the tracing layer. This shouldn't
happen, but the whole problem can be prevented altogether if the tracing
overlay isn't even used to lookup the radius. Thanks to Nadiya Sharifi for
bringing this to our attention.