Add '50' to default list of table page length options
This setting is used by many tables to provide options for possible page
lengths if pagination is used. Some tables use 50 as their default
value,  though it was not part of the default list of possible options.
This made the page length drop down control display an 'empty' value.
This is now fixed with 50 being part of the default list. Since only the
default changed, no setting migration is needed.