Fix loading of stack groups from URL
Moving the current project to the specified location didn't wait for the stack
group being loaded. This caused the project to not be initialized yet at this
point in time. Correctly returning promises from the internal stack loading
function fixes this, because moving the project can now wait until all
(dynamically) queued stacks and with them the project are loaded. Thanks to
@davidhildebrand for pointing out this bug.