Connectivity matrix: show warnings only once if both sources are loaded
If sources are loaded that either have no skeletons or above a defined
limit, a warning is shown. If both row and column sources are loaded in
the Connectivity Matrix, so far one warning was shown for each source
(in case there was something to warn about). This changed now so that
only one warning or confirmation is displayed.

Also, the skeleton source manager's loading methods now accept a
parameter for silent operation.

Fixes catmaid/CATMAID#1387