Tracing layer: fix occasional rounding error when moving with [ or ] between VNs
Using a virtual node's location as the basis for calculating a new virtual
node's location can cause the accumulating of floating point rounding errors.
This can be prevented by always using real nodes to determine a new virtual
nodes position. This was sometimes noticeable when using the brackets to move
along an edge.