Stack viewer: add flag to allows update while invisible
If a layer is invisible, it is not updated on redraws (e.g. on a
location change). This behavior can now be changed if a layer sets the
'updatedHidden' flag. It will make the layer update on a redraw() call
even though it is hidden. This is needed for instance to allow updating
tracing data while navigating along a skeleton with all nodes being
hidden.