3D viewer: don't show empty user colormap dialog
The user colormap dialog show only information if a user-based coloring
method is selected. If this is not the case, an empty dialog was shown
before. Now a warning that no user-based coloring method is selected is
shown. Also, a title attribute is added to the "User colormap" button to
explain this requirement.