Review widget: keep state when finishing review segment
The UI update change introduced in c4d973be caused the disposal of the
currrent state, i.e. which segment and node was reviewed last. This in
turn caused a workflow interruption if a segment was finished: the E
key as well as Q and W didn't work anymore. This is fixed by not ending
the current review when finishing a segment.