Connector table: act as skeleton source and update back-end API
This change replaces many internals of the connector table. The widget now
registers a skeleton source which includes the result skeletons. The most recent
Datatables API is used for better readability and consistency with the rest of
the code base. This in turn lead to an update of the connector list API, which
now also uses more readable parameter names that are more consistent with other
endpoints. Tests have been updated and the API documentation added.