Add new API for compact-skeleton endpoint, deprecate old one
The old API is still available but is marked as deprecated and will be
removed in a future version. There is no measurable difference in
performance (onlt the parameterization changed). The new endpoint is:

  GET /{project_id}/skeletons/{skeleton_id}/compact-detail

which returns the exact same data, but is parameterized with plain GET
parameters. The returned data structure is explained in the API comments
and not represented as Swagger type, because Swagger currently doesn't
support nested array definitions.