Importer: support ontology specification in project.yaml files
Project objects defined in a project.yaml file or through the internal
API can now provide a field named "ontology". It keeps a list of root
node objects, each with the fields "class" and "children". Class gives
the ontology class name and children is a list of child nodes. Each node
below the root has a "class", a "children" and a "relation" fields. The
relation field is a relation name that links the child to the parent.
For instance:

project:
   name: "test"
   ontology:
     - class: 'Metazoa'
       children:
         - relation: 'has_a'
           class: 'Protostomia'
           children:
             - relation: 'has_a'
               class: 'Lophotrochozoa'