Tracing overlay: fix missing promise resolution in move to child
Pressing the closing bracket ("]") key triggers a move to the current
node's child (cycle if more than one). This stopped working after trying
to got to the child of an end node. The source of the problem was a
missing promise resolution if the no move could performed (because there
are no children). This is fixed now.

Also add a reject call in case cycling fails, which could result in the
same behavior (a pending promise).

Thanks to Steven Arthur Calle-Schuler for reporting this bug!