3D viewer: allow picking of arbitrary points on skeletons
This is done by color mapping skeletons and finding the clicked skeleton by
looking up the color information in a skeleton map. With the knowledge of what
skeleton was clicked, the X,Y,Z coordinate of each rendered fragment is
rendered separately to a texture and retrieved from JavaScript. This information
is then used to find the closest (virtual) node to the location of interest.
The view will be moved to the given location and the node will be selected.

This had to be done in a rather cumbersome way to save memory by reusing the
same texture for X, Y and Z as well as to not loose precision by using the depth
buffer, which isn't as flexible and predictable with WebGL. 32 bit floating
point location components are packed into RGBA values.