Stack: remove special behavior for tile source 2
Tile source type 2 had special behavior for interacting with a
proof-of-concept Volumina tile server. This would have the CATMAID
backend act as a proxy for stack listing and included a URL for
uploading segmentation labels.

This behavior was never fully functional or deployed and conflicts with
planned refactorings, so is being removed.

If re-implemented, it would be better follow the pattern of render
service and DVID interop, i.e., a middleware for dynamic listing of
stacks and dedicated tile source types for static stacks.