3D viewer: respect model visibility when removing connector restriction
Removing a connector restriction now correctly sets pre/post visibility of
skeletons based on the corresponding skeleton model fields. Before, all
connectors were shown all the time.

A DOM checkbox update has been removed as well, because it was not functional
anymore. In the past this likely was used to update the selection table. For
connector restrictions this doesn't work very well in the first place: A
skeleton can have some pre-synaptic connectors (and others of course) shown and
some not, which can not be represented with a meaningful checkbox state.

Fixes catmaid/CATMAID#1426