Prepare Postgres 9.6 compatibility
The migration in this new pgcompat Django app adds a compatibility
function that allows us to use the Postgres 9.6 type signature of the
to_regclass() function.  Before Postgres 9.6, to_regclas() accepted one
cstring parameter. With Postgres 9.6 this becomes a text parameter. The
added function will *not* prevent updates from Postgres 9.5 to 9.6
through pg_upgrade or via pg_dump (tested). While the error is logged,
but ignored.

Updating actual CATMAID functions that use to_regclass() will be done
separately.

The introduction of this function is done as a separate app, because
Django allows the introduction of new dependencies to the initial
migration if they are separate apps. Using another initial migration
does not work. This allows current Postgres 9.5 setups to first add the
compatibility function and only then make changes to the existing
functions.

The pgcompat Django app and references to it can be removed once we
require Postgres 9.6.