Support Postgres 9.6
All uses of the to_regclass() function had to be updated due to
signature change of this function in Postgres 9.6. Postgres 9.5,
however, is still version listed in our requirements. So to still
support it, a compatibility function has been added in 91f05d7a7. It
allows Postgres 9.5 queries and functions to use the Postgres 9.6
version of to_regclass(). To make sure this is the case for all of the
CATMAID Django app, this compatibility function is added as a separate
app, listed as a dependency for the CATMAID app. Updates from Postgres
9.5 to Postgres 9.6 should still work without problems.