Tag table: have source manage update other widgets on init
Unfortunately, we still have to trigger an update of other skeleton
source pull downs manually. This should be taken care of by widget
registration or events (like it is done with unregistering a source).

Also make sure the actual skeleton source is destroyed when closing the
widget. Otherwise it would still show up in pull down menus of other
widgets.