Connectivity widget: store skeleton models instead of only names
This makes it easier to keep track of the original color of input
skeletons. This can now also be actively enabled by a checkbox. This is
desabled by default, so that the original behavior is kept. If enabled,
pulling (or subscriber) widgets will receive the original colors of the
skeletons as they were added to the source widget. This will make it
possible to use defined skeleton model colors also in the connectivity
Graph Plot.

See catmaid/CATMAID#757