Remove h5py dependency
In the version we ued (2.5.0), h5py required to be built during
virtualenv update, which increased setup time. While this is not needed
anymore with 2.6.0, we didn't really use h5py except for tile loading in
tile source type 3 and for mesh loading. Tile source type 3 has to be
explicitely enabled now and a note has been added to the documentation.
And since meshes have been replaced with volumes, there is no need
anymore to have this library listed as a dependency.

See catmaid/CATMAID#1339