Review widget: add note that virtual nodes are not included in count
The displayed number of nodes doesn't match the number of review
hops/steps, because virtual nodes are traversed as well, but they are
not considered for the count. This lead to some confusion among tracers
To better explain what this number means, a note has been added to the
widget UI. While a better solution would maybe show the number of
virtual nodes instead (maybe even for all open stacks), this should
reduce some confusion for now.

Also show/hide complete widget content with start/stop of review. This
makes it easier to show and hide the added note.