UI: propagate key event to active widget and add support for keyup event
In case an event is not handled by the UI instance, the tool and the
project, the active widget is now asked as well to handle a key event.

Additionally support for keyup events has been added. If a project, tool
or widget implements the handleKeyUp(event) function, it will be called
on key release. For the keydown event the handleKeyPress(event) is used.