Statistics widget: fix timezone conversion corner case
Timezone definitions change over time, which makes it important to use
the correct timezone offset for a particular date. Postgres was doing
this correctly when querying statistics for a particular time range. The
conversion back to the user/request time zone, didn't respect these
changes and only used the original definition. This caused problems in
some corner cases: For instance, EST (America/New_York) was originally
defined with an UTC offset of -4:56 (in 1883). Since 1920 this has been
updated to -5:00. The way timezones were applied to the query time range
caused the original offset to be in use. So everytime a timestamp was
done four minutes to midnight on the day before the start date of the
query time range, a conversion to UTC would lead to an unexpected date
(the day before the start date) and the the statistics back-end would
raise an error. This is now fixed by using pytz's localize() function
for creating a timezone aware start date end and end date.

Fixes catmaid/CATMAID#1425