Statistics: make timezone fix work work empty start/end date
If start date or end date are not provided by the request, the back-end
will use a date ten days back from the end as start and the current
time as end date. If this is the case, the datetime object will already
have timezone information attached, which causes an error if localize()
is used. Therefore, localize() is now only used to attach timezone
information to dates provided by the request.