Synapse fractions: allow manual color changes through graph legend
Clicking on any of the individual color swatches in the graph legend
will open a new color picker, which allows modification of the
respective group/partner color.

Unfortunately, our color picker can't be used very flexible. This is why
a separate color button had to be created which is moved on top of a
color swatch in the legend if the latter is clicked.

Fixes catmaid/CATMAID#1319