Project: add stack viewer to nearest CMW root node
Previously stack viewers were always added to the CATMAID root window. In order to use them within a widget, they need to be added to a different CMWRootNode: with this change the stackviewer finds its ancestor root node, and only if it cannot find one defaults to the CATMAID root window.