Stack Viewer: Can be created in non-root window
Previously, the stack viewer could only live in a new CMWWindow which it creates itself. In order to add a stack viewer to an existing widget, it needs to be added to an existing CMWWindow: this change allows that window to be passed in as an argument in the constructor.