Stack viewer: Make scale bar visibility accessible
Previously, the scale bar (slice benchmark)'s visibility was not controlled by the stack viewer and was not programmatically accessible. In order for other code to control the scale bar visibility, it had to go digging through the layer control elements and manually trigger the checkbox. This adds a boolean to the stack-viewer which allows other code to manage the visibility.