Overlay: prevent queue clashes w/ stack viewers
Previously, when adding requests to the queue, they would be given an ID based on the primary stack. When multiple stack viewers are used, this caused requests from one stack to overwrite other stacks' existing requests in the queue. This commit gives the request an ID based on the stack by which the request was made.