Stop UI key actions from executing twice
UI actions are currently Undo and Open-Widget-Dialog and it was possible
to execute them twice with a single key shortcut. Since f19b7d9 internal
key events are also generated for key-up events. A different handler is
used in this case and for all typical handlers (project, tool, widgets)
the correct handler is selected. UI actions however were executed for
both key-up and key-down. This is fixed now by following the naming
convention and also only execute UI's handleKeyPressAction() on
key-down.

Thanks to @aschampion for the pointer on how to reproduce this (undo
executed twice).

Fixes catmaid/CATMAID#1456