Window maker: don't require widgets to provide ID for controls/content
A widget configuration contains functions to create controls and content
for a particular widgets. If a createControls() method was provided, the
Window maker also required a controlsID field. A content ID was also
always mandatory. There is however now real need to enforce this and
many configurations don't refer to these IDs anywhere else and wouldn't
need to provide them in the first place.

Widgets now don't need to provide any of the two ID fields (controlsID
and contentID) anymore. They are set if available and otherwise the
control panel and/or the content container will have no ID.