Connector viewer: don't show result skeleton source in drop-down
A separe BasicSkeletonSource instance is used to manage the result data
set of available skeletons. This source was displayed in the widget's
source drop down, due to not being excluded explicitly. This has been
fixed and now neither the input source nor the result source show up.

Fixes catmaid/CATMAID#1467