WebGL contexts: Add checks and error dialog
Previously, widgets which create webGL contexts do not check to make sure that they are not going to hit the browser's limit. Now there is a constant containing the limit, a method to check the number currently in use by CATMAID, and a specialised error dialogue which context-creating widgets should raise if there are already too many contexts.