Improve performance of spatial queries
For each of the following tables two new indices are created:
treenode_edge, treenode_connector_edge and connector_geom. The fist
index is a Z min/max range index, which makes it fast to filter on Z
ranges. This is an expression index, so no actual min/max computations
have to be made. The second index is a 2D GIST index, which allows for
fast XY bounding box queries. Thanks to @Remi-Ci's help on the PostGIS
mailing list, I now know that both work very well together.

This makes node queries more than a magnitude faster: A test query
with 6799 treenodes went down from 600ms to 34ms and query returning 309
connectors went down from 50ms to 4ms. In the initial implementation,
the returned query results are exactly the same to verify validity. This
was however changes and now only nodes with in the range [z1, z2) are
queries, i.e. the next section is exclusive. Since commit 4c52ce77ba
coordinates transformed from project space into space space, it is
always rounded down to the last section. This makes it possible to
exclude nodes on the next section when querying nodes.

So far I didn't test insertion performance, but I don't think this will
be big problem. The ND index is also still required for the
ST_3DWithin() false positive removal. But as soon this is replaced, we
shouldn't need the ND index anymore (which could help insertion
performance, if it really would be a problem).

This update also improves access to orthogonal fields of view by about
the same amount.