Profiling middleware: return profile with response
This makes it much quicker to retrieve profiles for individual
endpoints, because no separate file transfer or remote analysis has to
be performed. Profiles are generated if the middleware is active and the
request parameters contain the 'profile' field. Returned is now a JSON
response which provides the original response in its content field and
the profile in a profile field. Writing to file on disk is now an
optional operation, which can be activated by using a field named
'profile-to-disk' as part of the request. Note, the profiling middleware
has to be activated manually.