3D viewer: make custom tag regex case-insensitive
The tag matching for 'uncertain' and 'todo' is already case-insensitive
by making the tag to compare to lowercase. This made the custom regex
fail if it contained upper case letters. To fix this, the custom tag
regex is now made case-insensitive.

Thanks to Steven Calle-Schuler for bringing this to our attention.