Graph widget: make SVG export work around bugs in Illustrator
Since Adobe doesn't seem to be able to implement a standard conforming
SVG importer, I had to come up with some hacks to make it accept line
markers (arrow heads). Illustrator reads line markers, but then won't
display the line they are attached to (while all other viewers/editors
show it). Therefore, a line with a marker is now duplicated. The copy is
hidden, so that only the marker is shown.