Layer controls: make layer removal button remove correct layer
All remove handlers shared the same closure and therefore referenced the
same variable that kept track of the layer key for the respective
button. This has been fixed by moving the remove handler creation into a
separate function, which provides a new closure for each callback.

Thanks to @davidhildebrand for bringing this to our attention.