Tag Table: persist sort and filter across refresh
Resolves #1482

Previously, all sorting information and filters (including the skeleton filter) were deleted when the Refresh button was clicked.

Now, the constraint skeleton source is not cleared, and filter and sort information is persisted (although this does not guarantee identical order due to in/del changes). The page is still reset, as is the label selection.

Also, renamed spurious footer labels.