Annotation front-end model: only send ann. removal event on actual removal
The EVENT_ANNOTATION_DELETED event is now only triggered if the
respective annotation isn't used at all anymore and was therefore
removed from the system. Before, it was triggered also for
de-annotations, even if the annotation was still in use on other
neurons. This caused the annotation cache to drop information on such an
annotation, which in turn broke annotation auto completion. Only after a
cache refresh, such annotations would show up again.

Fixes catmaid/CATMAID#1503