Tile layers: don't request broken section tiles
So far, if a stack is configured to hide the current section rather than
skip to the closest available one, tiles have still been configured so
that they would trigger actual image data requests---at for redraw()
calls that originate in a resize event. This caused 404 errors plus CORS
errors if WebGL is used. This is fixed now by making all tiles invisible
if they are on a broken section and should be hidden.

Fixes catmaid/CATMAID#1506