Data views: use mirror index and not position for opening links
A mirror position should only be used in comparison to other mirror
positions. Before templates didn't respect that and used the actual
position value as mirror index on the front-end. This has been updated
and now only indices into the list of mirrors ordered by their position
are used. This makes data views independent of the actual position
values.