Tile layer: fix selection of fastest stack for 0 and 1 mirrors
Thanks to @clbarnes for noting this problem. A reduce(f) call without
initial value will raise a TypeError when presented with an empty list
is passed and for a one-element list it will return this element without
calling f. With an initial value the intial value is returned for an
empty list (null in this case) and the f is run for a single element.

Fixes catmaid/CATMAID#1513
See catmaid/CATMAID#1514