Circuit graph: plot: add missing removeSkeletons() function
Skeleton sources are expected to have this function. Without it, the
removal of a loaded skeleton can cause errors. This would for instance
happen if a skeleton that is appended to the Circuit Graph Plot is
merged into another skeleton. The merge causes the skeleton source
manager to try and replace the loaded model with the merge result.