3D viewer: fix UVs for bottom row tiles in Z plane with images
The bottom row of Z plane with images used to have a wrong UV
coordinates, which was visible by not rendering the maximum zoom level
("max") correctly: The lower-right triangle on which part of the image
data was mapped wouldn't display correctly. The V component of its last
vertex was wrong---it always has to be 1, because tiles are aligned with
the upper boundary of the bounding box, clamping only happens at the
bottom and the right side.